Factor out convert-doc.js and policy-schema.js.
